+++ src/app/components/Metadata/index.test.jsx
@@ -2,6 +2,20 @@ import React from 'react';
 import Metadata from './index';
 import { shouldShallowMatchSnapshot } from '../../../testHelpers';
 
+const iconSizes = {
+  'apple-touch-icon': [
+    '72x72',
+    '96x96',
+    '128x128',
+    '144x144',
+    '152x152',
+    '192x192',
+    '384x384',
+    '512x512',
+  ],
+  icon: ['72x72', '96x96', '192x192'],
+};
+
 const metadataSnapshotTest = (
   testDescription,
   isAmp,
@@ -63,7 +77,7 @@ const metadataSnapshotTest = (
 
     shouldShallowMatchSnapshot(
       'should render correctly',
-      <Metadata {...metadataProps} />,
+      <Metadata {...metadataProps} iconSizes={iconSizes} />,
     );
   });
