+++ src/app/containers/Metadata/index.test.jsx
@@ -70,6 +70,7 @@ const metadataProps = (
   type,
   service,
   showArticleTags,
+  iconSizes: null,
 });
 
 const linkedDataProps = (
